fix: make all migrations idempotent and production-ready
- Migration #3: IF NOT EXISTS for oil_balance column, DROP POLICY IF EXISTS
- Migration #4: IF NOT EXISTS on CREATE TABLE, FK on user_id, idempotent policies
- Migration #5: Idempotent unique constraint check, DROP TRIGGER IF EXISTS
- Migration #11: Exception handler for forward reference to system_config
- Migration #12: IF NOT EXISTS guard for linked_wallet_address column
- Migration #15: DROP POLICY IF EXISTS for referrer wallet view
- Migration #17: IF NOT EXISTS on both tables, FK on user_id, idempotent policies
- Migration #18: IF NOT EXISTS on table, FK on user_id, idempotent policies/indexes
- Migration #19: DROP POLICY IF EXISTS for bonus_claims insert
- Migration #20: Check both constraint names (auto-generated and explicit)
- Migration #23: Remove redundant PK index, DROP POLICY IF EXISTS
- Migration #24: DROP POLICY IF EXISTS for wallet update policy
- Migration #25: Fix DECIMAL to INTEGER type mismatch in increment_task_progress

+++ supabase/migrations/20260204000005_add_task_progress_rpc.sql
@@ -3,12 +3,12 @@ CREATE OR REPLACE FUNCTION increment_task_progress(
   p_user_id UUID,
   p_task_id TEXT,
   p_date DATE,
-  p_increment DECIMAL,
-  p_target DECIMAL
+  p_increment INTEGER,
+  p_target INTEGER
 ) RETURNS VOID AS $$
 DECLARE
-  v_current_progress DECIMAL;
-  v_new_progress DECIMAL;
+  v_current_progress INTEGER;
+  v_new_progress INTEGER;
   v_completed BOOLEAN;
 BEGIN
   -- Insert or get current progress
